+++ ax101/code/course-project/src/production-order-system/view/BatchList.tsx
@@ -10,19 +10,22 @@ export const BatchList = ({ allBatches, onFreeze }: BatchListComponentProps): JS
     <table className="table">
       <thead>
         <tr>
-          <th scope="col">#</th>
-          <th scope="col">State</th>
-          <th scope="col">Product</th>
+          <th scope="col">Batch Number</th>
+          <th scope="col">Size</th>
+          <th scope="col">available material</th>
+          <th scope="col">used by</th>
+          <th scope="col">used for order</th>
+          <th scope="col"></th>
         </tr>
       </thead>
       <tbody>
         {allBatches.map(({ batchId, batchSize, availableMaterial, machines, orders }) => (
           <tr key={batchId}>
-            <th scope="row">{batchId}</th>
+            <th scope="row">{batchId.substr(0, 13)}</th>
             <td>{batchSize}</td>
             <td>{availableMaterial}</td>
-            <td>{machines}</td>
-            <td>{orders}</td>
+            <td>{machines.join(', ')}</td>
+            <td>{orders.join(', ')}</td>
             <td>
               <button onClick={() => onFreeze(batchId)}>freeze</button>
             </td>
